Data acquisition + Classes update
